// src/enums/event-type.enum.ts
export enum EventType {
    WORKSPACE_CREATED = 'WORKSPACE_CREATED',
    WORKSPACE_UPDATED = 'WORKSPACE_UPDATED',
    WORKSPACE_ARCHIVED = 'WORKSPACE_ARCHIVED',
    WORKSPACE_REOPENED = 'WORKSPACE_REOPENED',
    WORKSPACE_MEMBER_ADDED = 'WORKSPACE_MEMBER_ADDED',
    WORKSPACE_MEMBER_REMOVED = 'WORKSPACE_MEMBER_REMOVED',
    WORKSPACE_MEMBER_JOINED = 'WORKSPACE_MEMBER_JOINED',
    WORKSPACE_DELETE = 'WORKSPACE_DELETED',

    CARD_CREATED = "CARD_CREATED",
    CARD_UPDATED = "CARD_UPDATED",
    CARD_MOVED = "CARD_MOVED",
    CARD_REORDERED = "CARD_REORDERED",
    CARD_ARCHIVED = "CARD_ARCHIVED",
    CARD_RESTORED = "CARD_RESTORED",
    CARD_DELETED = "CARD_DELETED",
    CARD_DUPLICATED = "CARD_DUPLICATED",
    CARD_ATTACHMENT_ADDED = "CARD_ATTACHMENT_ADDED",
    CARD_ATTACHMENT_REMOVED = "CARD_ATTACHMENT_REMOVED",
    CARD_MEMBER_ASSIGNED = "CARD_MEMBER_ASSIGNED",
    CARD_MEMBER_REMOVED = "CARD_MEMBER_REMOVED",

    LIST_CREATED = "LIST_CREATED",
    LIST_UPDATED = "LIST_UPDATED",
    LIST_ARCHIVED = "LIST_ARCHIVED",
    LIST_RESTORED = "LIST_RESTORED",
    LIST_DELETED = "LIST_DELETED",
    LIST_MOVED = "LIST_MOVED",
    LIST_REORDERED = "LIST_REORDERED",
    LIST_DUPLICATED = "LIST_DUPLICATED",

    BOARD_CREATED = "BOARD_CREATED",
    BOARD_UPDATED = "BOARD_UPDATED",
    BOARD_ARCHIVED = "BOARD_ARCHIVED",
    BOARD_RESTORED = "BOARD_RESTORED",
    BOARD_DELETED = "BOARD_DELETED",
    BOARD_MEMBER_ADDED = "BOARD_MEMBER_ADDED",
    BOARD_MEMBER_REMOVED = "BOARD_MEMBER_REMOVED",
    BOARD_OWNER_CHANGED = "BOARD_OWNER_CHANGED",

    COMMENT_CREATED = "COMMENT_CREATED",
    COMMENT_UPDATED = "COMMENT_UPDATED",
    COMMENT_DELETED = "COMMENT_DELETED",
}